+++ src/main/java/org/iimsa/company_service/domain/repository/CompanyQueryRepository.java
@@ -1,6 +1,7 @@
 package org.iimsa.company_service.domain.repository;
 
 import org.iimsa.company_service.domain.model.Company;
+import org.iimsa.company_service.domain.model.CompanyId;
 import org.iimsa.company_service.domain.query.CompanyQueryDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
