+++ src/main/java/org/iimsa/company_service/domain/service/dto/CompanyManagerResponse.java
@@ -1,4 +1,4 @@
-package org.iimsa.company_service.infrastructure.client.dto;
+package org.iimsa.company_service.domain.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
