+++ src/main/java/org/iimsa/company_service/domain/service/dto/HubResponse.java
@@ -1,7 +1,8 @@
-package org.iimsa.company_service.infrastructure.client.dto;
+package org.iimsa.company_service.domain.service.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
 import java.util.UUID;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -9,4 +10,5 @@ public record HubResponse(
         @JsonProperty("hubId") UUID id,
         @JsonProperty("hubName") String name,
         @JsonProperty("address") String address
-) {}
+) {
+}
