+++ src/main/java/org/iimsa/company_service/infrastructure/client/CompanyManagerClient.java
@@ -1,10 +1,11 @@
 package org.iimsa.company_service.infrastructure.client;
 
-import java.util.UUID;
-import org.iimsa.company_service.infrastructure.client.dto.CompanyManagerResponse;
+import org.iimsa.company_service.domain.service.dto.CompanyManagerResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+
+import java.util.UUID;
 
 @FeignClient(name = "user-service")
 public interface CompanyManagerClient {
