+++ src/main/java/org/iimsa/company_service/infrastructure/client/HubClient.java
@@ -1,10 +1,11 @@
 package org.iimsa.company_service.infrastructure.client;
 
-import java.util.UUID;
-import org.iimsa.company_service.infrastructure.client.dto.HubResponse;
+import org.iimsa.company_service.domain.service.dto.HubResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+
+import java.util.UUID;
 
 @FeignClient(name = "hub-service")
 public interface HubClient {
