+++ src/main/java/org/iimsa/company_service/infrastructure/persistence/CompanyQueryRepositoryImpl.java
@@ -9,8 +9,8 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.iimsa.company_service.domain.model.Company;
+import org.iimsa.company_service.domain.model.CompanyId;
 import org.iimsa.company_service.domain.query.CompanyQueryDto.Search;
-import org.iimsa.company_service.domain.repository.CompanyId;
 import org.iimsa.company_service.domain.repository.CompanyQueryRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
